refactor(server/cache): extract caching logic to own file
